requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
fontawesomefree==6.4.2
nose==1.3.7
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uuid6==2023.5.2
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2==2.9.9
boto3==1.33.5
django-storages==1.14.2